[gcs-isolation] Add support for separate bucket in latest file.
Latests files are often used to track the latest built version of an
artifact that was archived. So far we have used the same bucket where
rest of the archives go as the house for latest files.

By nature this kind of file will be overwritten thus may violate
security policy for protected workloads. This CL adds option for
specifying a separate bucket for such files. Thoguh this would not be a
default behavior, secure workloads need to configure to use it to avoid
outages.

Initially we will support both uploads for an overlapping period before
consolidating two code paths (as mentioned in the TODO). This is
requirement to avoid/minimize downstream impacts.

R=jclinton@google.com,morawand@google.com

Bug:1262407,1296270
Change-Id: I7bbe9492cf9f5075d30c246ba57f009fa6bcb343
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3539564
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Wanda Mora <morawand@google.com>
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>