Add a presubmit to prevent unnecessary new uses of recipe configs.
This change adds a presubmit that prevents adding new recipe configs for
builders that are not connected to builders with existing recipe
configs. This should allow for adding a tester triggered by a builder
that already has a recipe config or a try builder mirroring builders
that already have recipe configs.

The PRESUBMIT script relies on the generate_groupings script, which
abstracts the process of invoking the chromium/builder_config_migration
recipe to produce JSON files containing information about the not-yet
migrated builders for a given project. The builders that are contained in
the migration file for a project are determined by JSON files in the
migration/scripts/filters directory.

The PRESUBMIT script ensures that the migration JSON files are kept
up-to-date and also prevents changes that would add new entries to the
files.

Bug: 868153
Change-Id: I11a3faf14db421e15969c9e62cbf5a2ad0630e37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3511117
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>