Add chromiumos_codesearch recipe
This recipe builds ChromiumOS and generates a KZIP file used to enable
cross-references in code search.

This recipe was originally written for chromiumos/infra/recipes. build
repo and the codesearch recipe module are not available as DEPS there.
Since this is now in build, a followup CL will dedupe code from the
recipe module.

Bug: 1284439
Change-Id: I44a55fd690eefad6e13275f12172bc06e824598f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3543473
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Josip Sokcevic <sokcevic@google.com>
Commit-Queue: Gavin Mak <gavinmak@google.com>