Simplify chromium_checkout test.
Currently, the chromium_checkout test relies on chromium_tests code and
configs present in the static DB. The code being tested only depends on
the clobber value of the passed in builder spec and none of the test
cases even cover that. Rewriting the test to not rely on
chromium_checkout's callers decouples it from changes of its callers and
the existence of specific builder configs.

Change-Id: Ie7b4b5e2b424cee291d8350708d4a74e735087d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3561643
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>