Enforce execution mode of builders in mirrors.
Currently, a tester can be specified as a mirrored builder and a builder
can be specified as a mirrored tester. This change raises an exception
if a mirror specifies a builder that doesn't have COMPILE_AND_TEST as
its execution mode or a tester that does have COMPILE_AND_TEST as its
execution mode.

Change-Id: If4f0d0fbf3ba4564fb759936c8278f56a1d8fc30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3561642
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>