Read more compilator build output props
Previously, the Orchestrator only looked for the
"swarming_trigger_properties" prop in the Compilator's build output. To
remove the src checkout on Orchestrator, the Orchestrator will need to
read a few more outputs from the Compilator:
- affected_files, deleted_files, got_revisions, and the
src_side_deps_digest

Actually using these props will be done in another CL.

Also reformat the output's affected_files paths to be relative to src so
it's in the same relative format that the rest of the chromium try
logic expects.

Bug: 1287228
Change-Id: Iba7fc87347471be0fc689ed1cd378703ca61f9d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3566217
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Stephanie Kim <kimstephanie@google.com>