Add a migration operation to chromium/builder_config_migration.
This uses the grouping computation already present in the
chromium/builder_config_migration to determine for a given builder all
of the builders that need to be migrated together and outputs the bits
of starlark code that need to be added to the builder definitions to
migrate the builder config src-side. An upcoming CL will add a script
that abstracts executing the recipe.

Bug: 868153
Change-Id: I5a28cdd10ca9b9a0ba39975ca82cd47e89bbb0a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3569975
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>