[flakiness] shard GetTestHistoryResult RPC call to hourly
We verify the tests we determine as new with ResultDB through the
GetTestHistoryResult RPC call. We check results for the last 3 hours,
and this helps ensure that we've got the latest information.

GetTestHistoryResult RPC call times out when we use a single query
that searches for the last 3 hours. Utilizing time ranges this way
for this RPC call isn't performant, so in the meantime we will
break this call down into 3 separate calls that search for an hour
window. We'll use the aggregated data to determine whether the
tests we've identified as new are indeed new.

Bug: 1313712
Change-Id: I3d86001a674a313e1c8754d32361310d05deee86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3579634
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>