Set TEST_ONLY when creating chromium config for parent.
In the lookup_builder_gn_args function called for testers, the parent
config is looked up in order to get the GN args used by the builder, but
without setting TEST_ONLY=true, it still runs the logic that validates
that the HOST_PLATFORM and TARGET_PLATFORM are compatible. This change
will set TEST_ONLY=true when making the config for the parent so that
the validation logic is disabled.

Bug: 1314422
Change-Id: I6dec6352b3becea5ef72c8848db7600c7ceb1164
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3584749
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Kuan Huang <kuanhuang@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>