Enable overriding the TEST_ONLY value when configuring the build.
The chromium config was recently updated with a TEST_ONLY value that
would disable some of the validation checks that are irrelevant for a
builder that is just triggering tests and not doing any compilation.
Being able to force the TEST_ONLY value to True will be useful for
builders like the finch smoke test builder. The finch smoke test builder
takes in information identifying the builder that it will run tests for
with different finch configs. In order to ensure all necessary files are
present (e.g. source side specs), it would be useful to be able to reuse
the configuration code with the builder config of the copied builder,
but this fails if the builder is not a tester and the target platform is
not compatible with linux. Being able to force it to be test-only will
allow reusing the config rather than trying to guess the appropriate
configs to manually apply as is being done right now.

Bug: 1315426
Change-Id: I4f9d148cbd4a8cfc988a6841b0367de3f310c3d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3584748
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>