Remove btolsch@chromium.org from recipes/recipes/openscreen.expected/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I74e1a1942ff37c56eeeaee4c552e2cdd19064b83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3589514
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>