[gcs-isolation]: Switch to primarily supporting latest bucket.
For BCID L2 we need to disallow overwrites in gs://chrome-unsigned.
To handle this migration an announcement was made regarding the
migration of gs://chrome-unsigned artifacts to gs://chrome-latest.
Support was provided for both buckets while teams were given time to
handle the migration if they depended on these artifacts. Support for
both buckets is now being discontinued and we will focus on using

gs://chrome-latest as the primary bucket for uploading and maintaining
the latest build artifacts.
Bug:1262407,1296270
Change-Id: I2296c5618e49c91db5ab4d49b04743a3e8a2aab6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3590959
Reviewed-by: Wanda Mora <morawand@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Akash Mukherjee <akashmukherjee@google.com>
Commit-Queue: Camilo Jacomet <cjacomet@google.com>