Correct target_platform for testers running on linux thin bots
For CI builder/tester splitted builders, the testers are normally not
responsible for compile and could run on a different host platform than
the builder. That the testers might report a wrong target_platform. This
CL specified the target_platform for these kind of builders to correct
the issue.

Bug: 1314422
Change-Id: Id0e92540060973c7053fc3d8d12c384167e86a0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3591675
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Kuan Huang <kuanhuang@chromium.org>