Update the migration operation to account for standalone try builders.
The migration operation in chromium/builder_config_migration currently
assumes that all try builders will need to specify mirrors when migrated
src-side, so for standalone try builders it outputs snippets containing
a mirror of the form "ci/<try-builder-name>". This change detects
mirrored builders that are the same as the builder to migrate to exclude
it from the mirrors list and only includes the mirrors attribute in the
output if there are builders to actually be mirrored.

Bug: 868153
Change-Id: I4b0f64ce15fd75b74452a2f99844e2fe7388eaac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3594247
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>