Fix cross-project checking for mirroring relationships.
The previous code was checking that all of the mirrors are in the same
project, but didn't check that the try builder was in the same project
as its mirrors. This was actually the case for v8_linux_blink_rel until
recently, which had linux-blink-rel-dummy as its mirror. This updates
the check to consider the try builder also.

Bug: 868153
Change-Id: If0f4022f3f239b8dfe33b30517a8586421d6edfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3597914
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>