Remove the set_output_commit field from BuilderSpec.
Milo has been updated to display the ref instead of the commit hash for
output commits where the ref is for a tag. Since there is no longer a
downside to setting the output commit, all builders are now setting the
output commit and the set_output_commit field is no longer necessary.

Change-Id: I7d88a3f734e39e886430babeafe824be3bc278bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3606977
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>