Make filter.suppress_analyze universally suppress analysis.
The current implementation of suppress_analyze sets an exclusion with
the regex f.*, which will only suppress analysis if there is a file
starting with foo. This clearly lines up with the default test data
returned by tryserver.get_files_affected_by_patch but would not suppress
analysis if that were overridden with a file that does not start with f
or if the CL to be applied is for a dependency repo whose path doesn't
start with f. This change makes it so that suppress_analyze will always
suppress analysis regardless of the affected files. As a consequence, it
is no longer necessary to take extra exclusions.

Change-Id: I233025dd61849a572f7ae4a981e9ef2a576c4b02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3606971
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>