Fix detection of existing groupings in src-side migration presubmit.
When the format of the groupings was modified to include blocker
information, the presubmit was not updated to get the list of builders
in a grouping via the 'builders' key. This means the presubmit was
looking to see if either 'builders' or 'blockers' was the name of a
builder in the old groupings instead of builder names. This prevents
renames from occurring, which might be necessary during the migration
when the wrong name is used within the recipe.

Bug: 868153
Change-Id: I8e1de265bf73a2e64a7c968c40f8d90c25321799
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3606380
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>