Exclude HOST_* chromium_config_kwargs when verifying builder config.
It doesn't make sense for the HOST_* values in the chromium config to be
set to anything other than what the default would be so setting them
isn't even supported src-side. However, if they are set, the verifier
currently fails because it does not appear in the builder config created
from the src-side properties. This excludes the HOST_* values from the
comparison so that recipe configs that specify them can be migrated.

Bug: 1288604
Change-Id: I99f9566724b771bc2a58e4dea9961f664a6263cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3615077
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>