Make linux_optional_gpu_tests_rel mirror its own trybot definition
By doing this, we cut-out the use of PROVIDE_TEST_SPEC in the
definitions of the trybot's mirrors.

This lets the recipe migration scripts generate starlark builder-spec
snippets for the trybot, which makes it easier to migrate the trybot
to src-side builder configs.

This needs crrev.com/c/3608992 to land before this can land. Led build
with both this and that CL applied:
https://ci.chromium.org/swarming/task/5a8f34b67036d410?server=chromium-swarm.appspot.com

Bug: 1317134
Change-Id: Iebba669724a5532d98ef23589f49b2d1a5ad594d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3614857
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>