Report exoneration reasons from recipes.
With crrev.com/c/3607770, a reason field was added to exonerations
in ResultDB, alongside with an enumeration of reasons. This allowed
the capture of a machine-readable exoneration reason, in addition to
the human-readable exoneration reason.

The machine-readable reason will be used in Weetbix to determine
which tests would have failed builds or CQ Runs were it not for
specific exoneration strategies. This will allow bugs to be filed
for tests failing on CQ regardless of if the failures were masked by
CQ exoneration.

BUG=1314194
TEST=./recipes.py test run

Change-Id: I1267c415f7d21fbca17ea63c71467185a32b3522
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3622039
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Auto-Submit: Patrick Meiring <meiring@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>