Fix exposing trigger properties
The code ensured that the command lines were archived, but didn't update
the local variable with the new in-memory model of the properties, which
meant the digest wasn't exposed.

https://ci.chromium.org/ui/p/chrome/builders/ci/linux-chrome/48033/overview
is an example of a build where the digest is not set correctly.

Recipe-Nontrivial-Roll: build_limited
Bug: 1315426
Change-Id: Ic4fe28aafad4918d450393889c7379058e12b494
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3630869
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Auto-Submit: Stephen Martinis <martiniss@chromium.org>