Make mac_optional_gpu_tests_rel mirror its own trybot definition
By doing this, we cut-out the use of PROVIDE_TEST_SPEC in the
definitions of the trybot's mirrors.

This lets the recipe migration scripts generate starlark builder-spec
snippets for the trybot, which makes it easier to migrate the trybot
to src-side builder configs.

This needs crrev.com/c/3622322 to land (and cherry-picked to all
relevant branches) before this can land. Led build with both this and
that CL applied:
https://ci.chromium.org/swarming/task/5aaf0e99c2073810?server=chromium-swarm.appspot.com

Bug: 1317134
Change-Id: I3ae8c831b41b93cfcb580bd18b437e357fa46e92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3621420
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>