Make win_optional_gpu_tests_rel mirror its own trybot definition
By doing this, we cut-out the use of PROVIDE_TEST_SPEC in the
definitions of the trybot's mirrors.

This lets the recipe migration scripts generate starlark builder-spec
snippets for the trybot, which makes it easier to migrate the trybot
to src-side builder configs.

This needs crrev.com/c/3636106 to land (and cherry-picked to all
relevant branches) before this can land. Led build with both this and
that CL applied:
https://ci.chromium.org/swarming/task/5ac36d9e7ce57f10?server=chromium-swarm.appspot.com

Bug: 1317286
Change-Id: Ic4a98fcbf400456aea1874a5285471a0ee478574
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3635622
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>