Update latest file to download from latest_upload.gcs_bucket
Now that the archive module can overwrite the latest files within
gs://chrome-latest, it went unnoticed that it was downloading the latest file from the old location: gs://chrome-unsigned, and comparing versions
against it. This was incorrectly updating the latest version
to the version stored in chrome-unsigned, which has not been updated for
a few weeks now. Now, the module will download the file from latest_upload.gcs_bucket if it is set.

Bug:b/229955731
Change-Id: Ie2ff057af9587d13445bf561dd881830cbfd5ca2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3640517
Commit-Queue: Wanda Mora <morawand@google.com>
Reviewed-by: Akash Mukherjee <akashmukherjee@google.com>