Make android_optional_gpu_tests_rel mirror its own trybot definition
By doing this, we cut-out the use of dummy builders in the
definitions of the trybot's mirrors.

This lets the recipe migration scripts generate starlark builder-spec
snippets for the trybot, which makes it easier to migrate the trybot
to src-side builder configs.

This needs crrev.com/c/3646268 to land (and cherry-picked to all
relevant branches) before this can land. Led build with both this and
that CL applied:
https://ci.chromium.org/swarming/task/5ad27ed4abe16a10?server=chromium-swarm.appspot.com

Bug: 1317105
Change-Id: I48e7b0f2aff6651aca1af1b47c550deb840519b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3645534
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>