Reland "Delete "chromium.memory:Linux ASan LSan Builder" and related"
This is a reland of commit 5f8a477ea367091f3631d3906ea437c0bb2a15c9

The original CL was reverted because the chromium/src side changes
were not cherry-picked to branches. That meant this config was not
safe to remove. I've cherry picked the changes to relevant branches,
so this should be safe to reland.

Original change's description:
> Delete "chromium.memory:Linux ASan LSan Builder" and related
>
> This builder, and related builders, have been migrated to chromium/src.
> See the linked bug.
>
> Bug: 1317131
> Change-Id: I2e7de24d570e4873b184d8db90340b9248797c59
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3645540
> Reviewed-by: Garrett Beaty <gbeaty@google.com>
> Commit-Queue: Stephen Martinis <martiniss@chromium.org>

Bug: 1317131
Change-Id: I234fa320b9d3580401fb9e8b5bdc8a85b42d1915
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3650968
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>