Don't set the -worker flag when running swarming collect.
The -worker flag has been removed from the current version of the
swarming client and a change was landed to the swarming module to update
the swarming client to a recent version. This change must be made in
order for rolls into the build recipe to resume.

Bug: 1327351
Change-Id: I0f46206dab5a3cb8bcc4eb1d74045573033ba1b4
Recipe-Nontrivial-Roll: build_limited
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3655754
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>