Add option to analyze kzip when syncing generated files
Not all generated files are referenced in codesearch. A kzip's
required_inputs contains all files that might be touched during the
course of a CompilationUnit's compilation. Add the option to analyze a
kzip and not sync files that aren't referenced in it.

For ChromiumOS, this should help reduce generated file size of chroot
and out dirs from ~90G to <1G.

Bug: 1284439
Change-Id: I457f685a7c410a466cf91cc7ac59c3a346f2956f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3657826
Reviewed-by: Josip Sokcevic <sokcevic@google.com>
Commit-Queue: Gavin Mak <gavinmak@google.com>