tricium_clang_tidy_orchestrator: use approved bots
In bug 1316616, capacity was allocated for the addition of these two
bots + win10-clang-tidy-rel. While we wait on win10 capacity, go ahead
and enable these.

This also removes a comment about wasted resources; the intent is to use
this recipe as a genuine orchestrator, which only has 2 cores available
to it and is (often) sourceless, which is far less wasteful than 8 & the
requirement of a warm chromium checkout. :)

Bug: 1153919, 1316616
Change-Id: I53332bacf708ffe5a581592833e334e9789334d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3685510
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Auto-Submit: George Burgess <gbiv@chromium.org>