Add support for specifying clusterfuzz archiving src-side.
The clusterfuzz archive functionality runs a resource script in the
archive module to determine the files to includes in the archive
depending on the platform. It was decided to continue using the resource
script and the clusterfuzz_archive method from the archive API rather
than force all builders currently using it to specify the archive moudle
properties.

This removes the presence of the cf_* fields as blockers for migrations.

Bug: 868153
Change-Id: I653c5ec0c4572cc83f4b85972c068b708ebff4b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3662021
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>