Remove information about clusterfuzz archiving blockers.
Support for clusterfuzz archiving was added to the src-side configs
since replacing it with the use of the archive module would have been a
considerable undertaking. Since the clusterfuzz archiving is supported,
specs that use it are no longer blocked on removing it.

Change-Id: Ifc5d47a8e00f18adaa6dc99e5382d9fdcca167b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3696506
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Wanda Mora <morawand@google.com>
Reviewed-by: Wanda Mora <morawand@google.com>