tricium_clang_tidy_orchestrator: add suggestion parsing logic
We have to dedup comments, so we try to sort them. In their current
form, they're only sortable if no suggestions are made by clang-tidy
(which is the common case). This CL teaches our orchestrator to parse
the suggestions in a way that makes them sortable.

Bug: 1336328
Change-Id: I76ba42adbb9b1cca3f8f9deb5965967ff4983de1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3705886
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>