tricium_clang_tidy_script: don't fall back to `all`
Existing invocations of this script will fall back to building `all` if
we fail to find one or more headers after building everything in the
same gn target as the header. This causes some builds >1hr, which hits
timeouts.

We can likely get very close to similar outputs in way less time (and
with way fewer timeouts) if we instead just build some of the targets
which depend upon the target which contains the header(s) in question.

To be clear, this optimization becomes more important as we branch out
to more and more platforms: many headers constrained to a handful of
platforms would trigger this problem. If any one bot hits this `all`
logic, the timeout hurts the linting experience for all of the other
bots that terminate promptly.

Bug: 1336385
Change-Id: I75a05c242e150fe80a017cdaae184ee7226fe22e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3707439
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>