Add skipping_coverage property to code_coverage and update compilator
The Orchestrator needs the affected files to run code coverage (the actual src files, not just the names). Since some CLs have a very large number of affected files and coverage is skipped for those cases, this CL includes the affected files in the archived src-side deps, only if coverage is not skipped.

If the len of files to instrument <= MAX_CANDIDATE_FILES (so coverage
is not skipped), the compilator **will** continue to isolate "all_test_binaries" but also include affected_files in the archived
src-side deps.

If the len of files to instrument is > MAX_CANDIDATE_FILES, a
code_coverage.skipping_coverage property will be set to True by the instrument() method. This also means that the orchestrator doesn't need code_coverage related src-side deps if the triggered compilator is skipping coverage. In that case, the compilator **will not** isolate "all_test_binaries" (build artifacts to run coverage) nor include affected files in the archived src-side deps. The compilator build will also output a `"skipping_coverage" = True` build output property. If the
orchestrator sees that output prop, it will skip all code_coverage
related calls.


Bug: 1287228
Change-Id: I44677e39a79cac946ce7bb9cfa6a1846c0fdd16f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3714699
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Prakhar Asthana <pasthana@google.com>
Commit-Queue: Stephanie Kim <kimstephanie@google.com>