scripts/common: fix deadlock in ProcessRead
If try block
https://source.chromium.org/chromium/chromium/tools/build/+/main:scripts/common/chromium_utils.py;l=972;drc=fe58615b5ead92a8214e1c47383a789a58184125
raises exceptions, |procs| is blocked when writing output to pipes
as there those are not read in that case.

Then |proc| doesn't terminate and wait in
https://source.chromium.org/chromium/chromium/tools/build/+/main:scripts/common/chromium_utils.py;l=1159;drc=fe58615b5ead92a8214e1c47383a789a58184125
is blocked too.

This CL fixes them by consuming output from proc for the case proc's
outputs are not read completely.

Tested with revert of https://crrev.com/c/3692483 and
$ vpython ./runtest.py --annotate=gtest ~/chromium/src/out/Release/base_unittests


Fixed: 1335777
Change-Id: If8a9c4a1849d38f71d414f4b4d1216562f700c2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3735987
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Takuto Ikuta <tikuta@chromium.org>