steps.py: Overwrite failure_on_exit() for experimental tests
Because an ExperimentalTest doesn't overwrite failure_on_exit() today,
calling it falls back to the parent class's implementation:
https://source.chromium.org/chromium/chromium/tools/build/+/main:recipes/recipe_modules/chromium_tests/steps.py;drc=76e0ff95cd3d52d3021c9bdf31388563ecb627c2;l=664

And because it falls back without calling _experimental_suffix() to
use the modified suffix that all ExperimentalTests prefer to use, it
fails to find the proper suffix in the _failure_on_exit_suffix_map map,
and so defaults to returning True.

This leads to the recipe considering experimental tests as invalid
despite their having no failed tests, eg:
http://screen/3k2jKsshazYqhpg

This bug doesn't have much of an impact since ExperimentalTests don't
impact the result of the build, but it'd still be nice to fix to keep
things consistent.

Bug: 1340060
Change-Id: Id0b98a916730e3a82722be6feebcd1a218fce6fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3741466
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>