Remove orchestrator assertion for compilator build status
The orchestrator was asserting that the triggered compilator build is finished before collecting the compilator swarming task. If a without patch compilator build was triggered, the compilator build would sometimes not be finished by the time the orchestrator is done rerunning an individual test. I've only seen this assertion error for this race condition, so I think it's safe to remove.

I've also increased the swarming task collect timeout to 4m to give any
without patch compilator builds an extra minute to complete.

Bug: 1340162
Change-Id: Id63b4eaf19931f245376eff8fd85b67d4649ef5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3750137
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Stephanie Kim <kimstephanie@google.com>