Remove chromium_android.zip_and_upload_build.
The zip_and_upload_build method in the chromium_android API is unused
except for the test of it. zip_and_upload_build zips up the entire
source tree and build outputs to be downloaded by a tester. The
preferred mechanism is to create isolates for tests that contain the
necessary files and have the tests be swarmed.

Change-Id: I9c89fd085c21393eca74c8b587434077b4f5b04a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3751421
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@google.com>