Add a TODO in summarize_findit_flakiness() about its removal
This reporting mechanism that FindIt currently uses is fragile and
unowned, and places an undue burden on the chromium recipe's internals.
ie: Tearing all this out would make maintenance of the recipe easier.

Ideally FindIt (either as a whole, or in parts) is being replaced with
Weetbix. So this reporting mechanism should no longer be needed after
the replacement. Add a TODO mentioning that in the relevant code.

Bug: 1243174
Change-Id: I83aba184038e1960800e16328250eac219dcb2e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3751685
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>
Reviewed-by: Stephanie Kim <kimstephanie@google.com>