[V8] Enable flake-bisect by failure regexp
This enables a failure_regexp property on flake-bisect jobs. When
the regexp is given, test failures only count as such if the regexp
also matches the overall test-framework output. Matching the regexp
in the output of each single test is a TODO marked in the code.

Using this regexp can make it easier to bisect errors that are masked
by errors with a different output.

This change also unifies the two result cases when collecting test
results (exceptional and non-exceptional) to simplify the code.

Bug: v8:13011
Change-Id: Ic55424607f90163cb5277123b8a892d500a18864
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3757803
Reviewed-by: Alexander Schulze <alexschulze@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>