tricium_clang_tidy_orchestrator: add a better failure message
It was brought up that this message isn't easily actionable. Emphasize
that clang-tidy should never block the CQ, and make it obvious to folks
how to get a green bot if they're not developing on a branch (...it's
seemingly unintended that tricium-clang-tidy runs on branches in the
first place).

Bug: 1343619
Change-Id: Id2a099e69b61ee465d20d9e343089dbec05ae9e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3759883
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Auto-Submit: George Burgess <gbiv@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>