tricium_clang_tidy: use `goma.jobs` Ninja jobs
We're explicitly setting -j150 in the configuration for these builders.
Goma only exposes that property under `jobs`; `recommended_jobs` is a
computed value that makes no attempt to examine the configuration we're
passing to the builder.

If we remove the builder configuration that specifies 150 goma jobs,
goma will fall back to the recommended job count.

Bug: None
Change-Id: I50a3f21e3e49c3f7e235fa8eb8c7bc1d8179bcf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3760890
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>