Move instantiation of ResultDB object into generator.
SkylabTestSpec.create currently extracts the resultdb keyword argument
and instantiates the ResultDB object for the spec. This is in contrast
to all other spec types that take a ResultDB instance that is
constructed by the calling generator. This change updates it to bring
SkylabTestSpec in alignment with the other spec types.

Change-Id: I8f82bd01fa40a7fc6868a30ab231463dbe5d883d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3770008
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>