Don't attempt to deapply patch on tryserver runhooks failure.
When the code was added to deapply patch on tryserver runhooks failures,
the deapply_patch had special logic for setting tags indicating
transient failures. This is no longer the case, so deapplying the patch
unnecessarily runs extra steps for no benefit and leads to confusing
results such as compile-only try builders (for which 'without patch'
steps are unexpected) appearing to run 'without patch' phase in
monitoring if the build is canceled during runhooks.

This change removes the special logic and moves the runhooks
implementation into prepare_checkout since it was not called anywhere
else except for the test of the runhooks function which no longer had
any applicable test cases.

Change-Id: Iae685dbdad44273dc169b9aebfabd315a0b2df50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3785502
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>