Don't infra fail archive steps if latest file doesn't exist.
The release dashboard determines if the archive operation failed based
on the status of the generic archive step. If the latest version file
isn't present in GCS then attempting to download it results in the
download step having an infra failure status, which causes the top-level
generic archive step to also have an infra failure despite the file not
existing being totally expected for a new builder.

This changes it so that the top-level generic archive step doesn't take
the worst status of contained steps, but the status of the last step.
This means that if an individual step has some failure status but the
recipe code prevents the exception from propgating and runs additional
steps that are success the top-level step will show up as success.

Change-Id: If72bfc9205cb9eecc98cb89c5b7c48c393ca1a39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3792776
Commit-Queue: Wanda Mora <morawand@google.com>
Reviewed-by: Wanda Mora <morawand@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>