Refactor TestOptions.
This change makes TestOptions an attrs class, which grants immutablity
in a more concise manner than manually defining properties and moves
functionality concerning the test options into a method on the type.

TestOptionsFlags, another attrs class, was introduced to encapsulate
flag support for the options in TestOptions. Instead of inheriting from
a TestArguments base class, the tests instead use an instance of
TestOptionsFlags. This reduces the amount of inheritance and mutability.

Change-Id: If4e1d1088f962c32be1582e5792fa37ad55e74dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3780852
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Gregory Guterman <guterman@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>