Remove legacy scripts.
There are a number of scripts in the repository that predate recipe
bundles. There are no .gitattributes files referring to them, so they
aren't included in the recipe bundle for the repo and therefore couldn't
actually be executed on a builder.

runtest.py was referring to archive_profiling_data.py in a bit of code
that was unused because it was guarded with a conditional based on the
name of the builder, restricting it to be executed only if the builder's
name was one of two builders that no longer exist.

Bug: 1008256
Change-Id: If963d6f646d901f0c8675b4a7b3ecf02332f7d8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3794946
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>