Remove uses of build.python from squashfs.
The build.python method presented a similar API to the python module,
which was deprecated, but added a wrapper script that set PYTHONPATH to
include locations containing common modules. Instead, python can be
invoked directly using the step module and so long as the common modules
are needed, they can be imported by using the *_pythonpath methods.

Bug: 1329377
Change-Id: I1ea37cf911d0373ab2425be937637b2eef791a76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3803565
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>