Remove uses of build.python from test_utils.
The build.python method presented a similar API to the python module,
which was deprecated, but added a wrapper script that set PYTHONPATH to
include locations containing common modules. Instead, python can be
invoked directly using the step module and so long as the common modules
are needed, they can be imported by using the *_pythonpath methods.

The archive_layout_tests_results_summary.py script was made a resource
script of the test_utils recipe module since that was the only recipe
module calling it.

Bug: 1329377, 1086694
Change-Id: If30123e67c4825786bb63fee7d64c511e0fb8129
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3799194
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>