Clean up DEPS file.
The DEPS file currently pulls in depot_tools, but the CIPD bundle for
depot_tools is not used by any PRESUBMIT checks and recipes get their
depot tools through the recipe deps, so the bundle being pulled in by
gclient is unused.

Additionally, a script that is only used as a runhook was moved to a new
hook-scripts directory to keep scripts run by builders and scripts run
as part of a gclient checkout separate.

Bug: 1008256
Change-Id: I78f0dae3d0aad34ee47719570bc695a24c7f24e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3795460
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>