Add a recipe module for polymorphic builders.
A polymorphic builder is a builder that builds and tests chromium on
behalf of other builders. This enables creating a single builder that
can perform chromium operations in contexts besides CI or CQ without
having to duplicate configuration for the target builders.

This module provides the interface between the polymorphic builder and
the builder that would trigger the polymorphic builder (since the
polymorphic builder isn't configured for a specific builder config, the
trigger needs to provide information to target the correct builder).

Bug: 1346396
Change-Id: Ic7a166a30d01a8deedc710d108cf6437e6b01577
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3838601
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>